backend/requirements.txt
aiosqlite==0.21.0
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.4.0
authx==1.4.3
billiard==4.2.3
blinker==1.9.0
celery==5.5.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contextlib2==21.6.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.3
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-decouple==2.1
django-environ==0.12.0
django-jazzmin==3.0.1
django-ninja==1.4.5
django-ninja-extra==0.30.2
django-ninja-jwt==5.4.0
django-qrcode==0.3
django_celery_results==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
dnspython==2.8.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
drf-yasg==1.21.5
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.121.0
Flask==3.1.2
greenlet==3.2.4
gunicorn==20.1.0
h11==0.16.0
idna==3.10
inflection==0.5.1
iniconfig==2.3.0
injector==0.22.0
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
Markdown==3.7
MarkupSafe==3.0.2
ninja==1.13.0
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.0
pytest==9.0.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.3
qrcode==8.2
redis==4.4.0
referencing==0.37.0
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.29.0
rsa==4.9.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.44
sqlparse==0.4.2
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
whitenoise==6.4.0
